\name{margin-class}
\title{margin class}
\docType{class}
\alias{margin-class}
\description{Class for margins types}
\section{Objects}{
  \describe{
    \item{marginExp}{exponential distribution}
    \item{marginFrechet}{Frechet distribution}
    \item{marginGPD}{GPD distribution}
    \item{marginGamma}{Gamma distribution}
    \item{marginLnorm}{log-normal distribution}
    \item{marginUnif}{uniform distribution}
    \item{marginWeibull}{Weibull distribution}
  }
}
\seealso{
  \code{\link{mubz.copula}} for a description of the distributions and their
  parameters.
}
