\name{margin}
\title{Margin distributions}
\alias{marginExp}
\alias{marginFrechet}
\alias{marginGPD}
\alias{marginGamma}
\alias{marginLnorm}
\alias{marginUnif}
\alias{marginWeibull}
\description{
  Margin distributions
}
\usage{
marginExp
marginFrechet
marginGPD
marginGamma
marginLnorm
marginUnif
marginWeibull
}
\value{
  A \code{"\linkS4class{margin}"} object.
}
\seealso{
  \code{\link{mubz.copula}} for a description of the distributions and their
  parameters.
}