\name{spatial-class}
\title{spatial class}
\docType{class}
\alias{spatial-class}
\description{Class for spatial models}
\section{Objects}{
  \describe{
    \item{spatialWhittleMatern}{for the Whittle Matern correlation model}
    \item{spatialCauchy}{for the Cauchy correlation model}
    \item{spatialPower}{for the power model}
    \item{spatialPowerexp}{for the power exponential model}
    \item{spatialBessel}{for the Bessel correlation model}
  }
}