\name{spatial}
\title{Spatial models}
\alias{spatialWhittleMatern}
\alias{spatialCauchy}
\alias{spatialPower}
\alias{spatialPowerexp}
\alias{spatialBessel}
\description{
  Spatial models
}
\usage{
spatialWhittleMatern
spatialCauchy
spatialPower
spatialPowerexp
spatialBessel
}
\value{
  A \code{"\linkS4class{spatial}"} object.
}
